requirements.txt
discord.py
python-dotenv
pydigitemp